requirements.txt
PyGithub
gitpython
docker